+++ tests/tests_model_manager/test_mm2.py
@@ -314,11 +314,12 @@ def test_delete_propagated_terms_from_parent(mm):
     q2 = """
     MATCH (c:Class)-[:HAS_CONTROLLED_TERM]->(term:Term)
     WHERE term.`Term Code`='term1t'
+    WITH * ORDER BY c.label
     RETURN collect([c.label]) as res
     """
 
     res2 = mm.query(q2)[0]['res']
-    assert res2 == [['class4'], ['class2'], ['Apple']]
+    assert res2 == [['Apple'], ['class2'], ['class4']]
 
     mm.delete_terms_of_parent_class([['class2','Apple']]) 
 
@@ -675,11 +676,11 @@ def test_create_subclass(mm):
     assert res1 == [['class1', 'class3'],['class2', 'Apple'], ['class4', 'class2']]
 
     res3 = mm.create_subclass([['class1', 'MISSING CLASS']])
-    assert res3 == []
+    assert res3 == [[]]
 
     #returns list of classes and subclasses - [['parent_class', 'child_class']...]
     res4 = mm.get_subclasses_where()
-    assert res4 == [{'parent':'class1', 'child':'class3'},{'parent':'class2', 'child':'Apple'}, {'parent':'class4', 'child':'class2'}]
+    assert res4 == [{'parent':'class1', 'child':'class3', 'conditions':None},{'parent':'class2', 'child':'Apple', 'conditions':None}, {'parent':'class4', 'child':'class2', 'conditions':None}]
     
     #propagates terms to parent class
     mm.propagate_terms_to_parent_class()
@@ -687,10 +688,20 @@ def test_create_subclass(mm):
     q2 = '''
     MATCH (c:Class)-[:HAS_CONTROLLED_TERM]->(term:Term)
     WHERE c.label = 'class4'
+    WITH * ORDER BY c.label, term.`Term Code`
     RETURN collect([c.label, term.`Term Code`]) as res
     '''
     res5 = mm.query(q2)[0]['res']
-    assert res5 == [['class4', 'term2t'], ['class4', 'term1t']]
+    assert res5 == [['class4', 'term1t'], ['class4', 'term2t']]
+
+    cond  =  [{
+            'EXISTS>': {'include': [{'Test' : {'uri': ['neo4j://graph.schema#Term/S980028/S91301']}}]},
+            'NOT EXISTS': {'exclude': [{'Ser': {'rdfs:label': ['Y']}}, 'Pop', 'Asta']}
+        }]
+
+    res6 = mm.create_subclass([['class1', 'class3', cond]])
+    assert res6 == [['class1', 'class3', '[{"EXISTS>":{"include":[{"Test":{"uri":["neo4j://graph.schema#Term/S980028/S91301"]}}]},"NOT EXISTS":{"exclude":[{"Ser":{"rdfs:label":["Y"]}},"Pop","Asta"]}}]']]
+
 
     #To check propagated_from terms property