+++ tests/tests_query_builders/test_qb2.py
@@ -116,7 +116,8 @@ class TestGenerateQueryBody:
 
 def test_split_out_optional(qbr: QueryBuilder):
     # #Test case #1
-    labels = ['Subject', 'Sex', 'Exposure**', 'Exposure Unit**', 'Exposure Laterality**']
+    labels = ['Subject', 'Sex', 'Exposure', 'Exposure Unit', 'Exposure Laterality']
+    labels_opt = ['Exposure', 'Exposure Unit', 'Exposure Laterality']
     rels = [
         {'from': pair[0], 'to': pair[1], 'optional': i > 0}
         for i, pair in enumerate([
@@ -128,8 +129,8 @@ def test_split_out_optional(qbr: QueryBuilder):
     ]
     res = qbr.split_out_optional(
         labels=labels,
-        rels=rels,
-        oclass_marker='**')
+        labels_opt=labels_opt,
+        rels=rels)
     # print(res)
     expected_res = [
         (
@@ -163,8 +164,8 @@ def test_split_out_optional(qbr: QueryBuilder):
     ]
     res = qbr.split_out_optional(
         labels=labels,
-        rels=rels,
-        oclass_marker='**')
+        labels_opt=labels_opt,
+        rels=rels)
     expected_res = [
         (
             ['Subject', 'Sex'],
@@ -181,7 +182,8 @@ def test_split_out_optional(qbr: QueryBuilder):
     assert res == expected_res
 
     # #Test case #3
-    labels = ['Subject', 'Sex', 'Exposure**', 'Exposure Unit**', 'Visit**', 'Vitals**', 'Vitals Unit**']
+    labels = ['Subject', 'Sex', 'Exposure', 'Exposure Unit', 'Visit', 'Vitals', 'Vitals Unit']
+    labels_opt = ['Exposure', 'Exposure Unit', 'Visit', 'Vitals', 'Vitals Unit']
     rels = [
         {'from': pair[0], 'to': pair[1], 'optional': i > 0}
         for i, pair in enumerate([
@@ -196,8 +198,8 @@ def test_split_out_optional(qbr: QueryBuilder):
     ]
     res = qbr.split_out_optional(
         labels=labels,
-        rels=rels,
-        oclass_marker='**')
+        labels_opt=labels_opt,
+        rels=rels)
     expected_res = [
         (
             ['Subject', 'Sex'],
@@ -371,6 +373,37 @@ def test_where_rel_2(qbr):
     # print("\n")
     # print(Cypher_list)
     # print(expected_cypher_list)
+    assert Cypher_list == expected_cypher_list
+    assert data_binding_dict == {}
+
+
+def test_where_rel_3(qbr):
+    test_map = {
+        'nobs': {
+            'EXISTS>': {'include': [{'Test' : {'uri': ['neo4j://graph.schema#Term/S980028/S91301']}}]},
+            'NOT EXISTS': {'exclude': [{'Ser': {'rdfs:label': ['Y']}}, 'Pop', 'Asta']}
+        }
+    }
+    (Cypher_list, data_binding_dict) = qbr.list_where_rel_conditions_per_dict(mp=test_map)
+    expected_cypher_list = [
+        "EXISTS {MATCH (`nobs`)-[]->(x) WHERE (x.`uri` in ['neo4j://graph.schema#Term/S980028/S91301'] AND x:`Test`)}",
+        "NOT EXISTS {MATCH (`nobs`)-[]-(x) WHERE NOT ((x.`rdfs:label` in ['Y'] AND x:`Ser`) OR x:`Pop` OR x:`Asta`)}"]
+    
+    assert Cypher_list == expected_cypher_list
+    assert data_binding_dict == {}
+    
+    
+def test_where_rel_4(qbr):
+    test_map = {
+        'nobs': {            
+            'EXISTS': {'include': [{'Ser': {'rdfs:label': ['Y', 'U']}}, {'Pop': {'value': ['Safety']}}, 'Asta']}
+        }
+    }
+    (Cypher_list, data_binding_dict) = qbr.list_where_rel_conditions_per_dict(mp=test_map)
+    expected_cypher_list = [        
+        "EXISTS {MATCH (`nobs`)-[]-(x) WHERE (x.`rdfs:label` in ['Y', 'U'] AND x:`Ser`) OR (x.`value` in ['Safety'] AND x:`Pop`) OR x:`Asta`}"]
+    print(Cypher_list)
+    print(expected_cypher_list)
     assert Cypher_list == expected_cypher_list
     assert data_binding_dict == {}