+++ CHANGELOG.md
@@ -1,5 +1,36 @@
 Change Log
 All notable changes to the tab2neo package will be documented in this file.
+[2.0.3.0]
+
+## Added
+
+- Enabled provision of conditions on SUBCLASS via ModelManager
+
+[2.0.2.0]
+
+## Added
+
+- Excluded the labels marked with !! from the WITH and RETURN parts of the query generated for get_data_generic()
+
+[2.0.1.0]
+
+## Added 
+
+- Extended list_where_rel_conditions_per_dict function in Query Builder
+  
+[2.0.0.0]
+
+## Added
+
+- Support Neo4j version 5.x
+
+[1.3.11.1]
+
+Bugfix:Changed the method name in json for test cases to run
+
+[1.3.11.0]
+
+Added indexes in data_load
 
 [1.3.10.1]